requirements.txt
click==8.0.3
psycopg2-binary==2.9.1
black==21.10b0
redis==4.1.1
